// mobile/gen/models/GetMe.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export const getMe200RoleEnum = {
  UTILIZADOR: "UTILIZADOR",
  ADMINISTRADOR: "ADMINISTRADOR",
  RECEPCIONISTA: "RECEPCIONISTA",
  "// change or add new roles": "// change or add new roles",
} as const;

export type GetMe200RoleEnumKey =
  (typeof getMe200RoleEnum)[keyof typeof getMe200RoleEnum];

export const getMe200ProvinceEnum = {
  MAPUTO: "MAPUTO",
  XAI_XAI: "XAI_XAI",
  INHAMBANE: "INHAMBANE",
  SOFALA: "SOFALA",
  MANICA: "MANICA",
  TETE: "TETE",
  ZAMBEZIA: "ZAMBEZIA",
  NAMPULA: "NAMPULA",
  NIASSA: "NIASSA",
  CABO_DELGADO: "CABO_DELGADO",
} as const;

export type GetMe200ProvinceEnumKey =
  (typeof getMe200ProvinceEnum)[keyof typeof getMe200ProvinceEnum];

/**
 * @description User information retrieved successfully
 */
export type GetMe200 = {
  /**
   * @type string
   */
  id: string;
  /**
   * @type string
   */
  email: string;
  /**
   * @default "UTILIZADOR"
   * @type string
   */
  role: GetMe200RoleEnumKey;
  /**
   * @type string | undefined
   */
  firstName?: string;
  /**
   * @type string | undefined
   */
  lastName?: string;
  /**
   * @type string | undefined
   */
  phone?: string;
  /**
   * @type string | undefined, date-time
   */
  birthDate?: string;
  /**
   * @type string | undefined
   */
  province?: GetMe200ProvinceEnumKey;
  /**
   * @type string | undefined
   */
  city?: string;
  /**
   * @type string | undefined
   */
  neighborhood?: string;
  /**
   * @type object | undefined
   */
  doctorAvailability?: object;
  /**
   * @type string | undefined
   */
  doctorSpecialtyId?: string;
  /**
   * @type string | undefined, date-time
   */
  passwordChangedAt?: string;
  /**
   * @type string | undefined, date-time
   */
  lastLoginAt?: string;
  /**
   * @default false
   * @type boolean
   */
  isSuperUser: boolean;
  /**
   * @default false
   * @type boolean
   */
  isStaff: boolean;
  /**
   * @type string | undefined, date-time
   */
  deletedSelfAccountAt?: string;
  /**
   * @default true
   * @type boolean
   */
  isActive: boolean;
  /**
   * @type string, date-time
   */
  createdAt: string;
  /**
   * @type string, date-time
   */
  updatedAt: string;
};

/**
 * @description Authentication required
 */
export type GetMe401 = any;

export type GetMeQueryResponse = GetMe200;

export type GetMeQuery = {
  Response: GetMe200;
  Errors: GetMe401;
};